refactor: extract NRIC/QR scanner to IdScanner
